change lactate type float to double and add test calcLactateDiff

+++ domain/src/main/java/sogyo/wesley/ipsportdata/domain/Analysor.java
@@ -3,15 +3,15 @@ package sogyo.wesley.ipsportdata.domain;
 public class Analysor implements IAnalysor {
     String name = "";
     int power;
-    float lactate_one;
-    float lactate_two;
-    float lt_diff;
+    double lactate_one;
+    double lactate_two;
+    double lt_diff;
 
     public Analysor(String name) {
         this.name = name;
     }
 
-    public Analysor(int power2, float lactate_one2, float lactate_two2) {
+    public Analysor(int power2, double lactate_one2, double lactate_two2) {
         this.power = power2;
         this.lactate_one = lactate_one2;
         this.lactate_two = lactate_two2;
@@ -29,16 +29,16 @@ public class Analysor implements IAnalysor {
     }
 
     @Override
-    public float getLactate_two() {
+    public double getLactate_two() {
         return lactate_two;
     }
 
     @Override
-    public float getLactate_one() {
+    public double getLactate_one() {
         return lactate_one;
     }
 
-    public float calcLactateDiff() {
+    public double calcLactateDiff() {
         lt_diff = lactate_one - lactate_two;
         return lt_diff;
     }
